Grammar
Associates each word that is used to build the Markov chain with it's part of speech, so sentences tend to be a bit more grammatically correct.
The previous two words are now used for the input to generate the next word, instead of the last 11 letters.
Since Heroku doesn't have the memory necessary to tag words with their part of speech, so that has been offloaded to my computer, where I will run WordBankCreator.py periodically, which uploads the word bank and starting letters to a Google Drive. The main script now downloads those files.